Rename OfferAcceptService -> Offers::AcceptService
